feat: import path overrides
This commit adds a configuration option to override the import path for
specific Go packages, in case they are not correctly found by the plugin.

My usecase is generating schemas for a type defined outside of the
github.com/gravitational/teleport/api/types package.
Without this change the plugin was just trying the following, which
obviously doesn't work:

```go
import (
	types "types"
	wrappers "wrappers"
)
```