fix!: pass message types by pointer
Where this is used in the Teleport Terraform provider, message types are
usually protoc-generated structs, which usually should not be copied and
are not generally safe to pass by value.
This commit updates Copy(Type)ToTerraform to accept the message argument
via a pointer.

Since this is a breaking change, and I figure we might as well attempt
to follow semver in this repo, I've bumped the version to v2.0.0.